// src/app/core/models/RecuperarSenhaDTO/recuperarSenha.ts
export enum SubRotasRecuperarSenha {
    identificacao = "identifique-se",
    informarCodigoEnviado = "informar-codigo-enviado",
    linkConfirmacao = "link-confirmacao",
    novaSenha = "nova-senha",
    validarCodigoDeAtivacao = "validar-codigo-de-ativacao",
    avisoComStepper = "cadastro-atualizado",
    cadastroInexistente = "cadastro-inexistente",
    pathAviso = "/login/recuperar-senha/",
    aviso = "aviso"
}